+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- Add video and document downloading xAPI events
+
 ### Removed
 
 - IE11 is no longer supported