+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Redirect participant to the player once jitsi conference left
 - Redirect instructor to the waiting jitsi page once conference left
+- Only a webinar using jitsi can be created
 
 ### Removed
 
@@ -37,7 +38,6 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 
 - Request initial MUC config before modifying it
-- Only a webinar using jitsi can be created
 
 ### Fixed