+++ CHANGELOG.md
@@ -24,11 +24,12 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Redirect instructor to the waiting jitsi page once conference left
 - Only a webinar using jitsi can be created
 - Hide timed text pane when video is not a VOD
-
-### Changed
-
 - On app load instructor is redirected to the dashboard when the video
   is a live
+
+### Fixed
+
+- Implement useTranscriptTimeSelector with videojs
 
 ### Removed